requirements.txt
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.16.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nodeenv==1.9.1
platformdirs==4.3.6
pre-commit==3.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.44.0
typer==0.15.4
typing_extensions==4.13.2
uvicorn==0.33.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==0.24.0
websockets==13.1
